requirements.txt
Pillow==10.2.0
httpx==0.27.0
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.4.2
colorthief==0.2.1
transformers==4.38.2
torch==2.2.1
stability-sdk==0.8.5 